+++ CHANGELOG.md
@@ -1273,3 +1273,4 @@
 - [2026-05-18 19:12] fix(test): resolve race condition during test initialization
 - [2026-05-18 20:16] test(cache): add unit tests for cache business logic
 - [2026-05-18 20:24] docs(i18n): update readme and API contracts for i18n
+- [2026-05-18 20:38] fix(state): resolve race condition during state initialization